Dev #186: initial refactoring for preprocess_toolbox consumption
