+++ icenet/data/network_dataset.py
