+++ icenet/plotting/data.py
@@ -11,7 +11,7 @@ from math import ceil
 
 from icenet.data.datasets.utils import get_decoder
 from icenet.data.cli import date_arg
-from icenet.data.dataset import IceNetDataSet
+from icenet.data.network_dataset import IceNetDataSet
 from icenet.utils import setup_logging
 
 import matplotlib.pyplot as plt
